+++ package.json
@@ -7,7 +7,7 @@
   },
   "scripts": {
     "build": "tsc",
-    "start": "node --optimize_for_size --max_old_space_size=460 ../build/app.js",
+    "start": "node --optimize_for_size --max_old_space_size=460 ./build/src/app.js",
     "client": "cd tsclient && npm run start && cd ..",
     "heroku-postbuild": "npm install && cd tsclient && npm install && npm run build && cd .. && npm run build "
   },
