+++ tsconfig.json
@@ -64,7 +64,7 @@
     "forceConsistentCasingInFileNames": true  /* Disallow inconsistently-cased references to the same file. */
   },
 
-  "exclude": [
-    "../tsclient",
-  ]
+  "include": [
+    "src/app.ts",
+  ],
 }
